+++ src/Bulk.php
@@ -209,7 +209,7 @@ class Bulk
             } elseif ($actionData instanceof Document) {
                 $this->addDocument($actionData, $opType);
             } else {
-                throw new \InvalidArgumentException('Data should be a Document, a Script or an array containing Documents and/or Scripts');
+                throw new InvalidException('Data should be a Document, a Script or an array containing Documents and/or Scripts');
             }
         }